% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bad_words}
\alias{bad_words}
\title{Bad words}
\format{A character vector with 1383 words.}
\usage{
bad_words
}
\description{
A list of words from Luis von Ahn's Research Group
\url{https://www.cs.cmu.edu/~biglou/resources/}. The list contains 1383
words/terms that could be found offensive. The list contains a lot of words
that most people don't find offensive, but it is a good start.
}
\keyword{datasets}
