% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offenser-package.R
\docType{package}
\name{offenseR-package}
\alias{offenseR-package}
\title{A package that checks for offensive words in texts and ducuments}
\description{
A package that checks for offensive words in texts and ducuments
}
