% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin.R
\name{offensive_addin}
\alias{offensive_addin}
\title{Checks for offensive words in selection}
\usage{
offensive_addin()
}
\description{
Call this function as an addin to check for offensive words in selection.
Example is shown at \url(https://github.com/EmilHvitfeldt/offensiveR).
}
