% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_files.R
\name{offensive_check_files}
\alias{offensive_check_files}
\alias{offensive_check_text}
\title{Offensive check}
\usage{
offensive_check_files(path, words_add = character(),
  words_ignore = character())

offensive_check_text(text, words_add = character(),
  words_ignore = character())
}
\arguments{
\item{path}{path to file or to offensive check}

\item{words_add}{Character. Words to look for.}

\item{words_ignore}{Character. Words to ignore.}

\item{text}{character vector with plain text}
}
\description{
Perform a check for offensive works on document files or plain text.
}
\examples{
# Example files
files <- list.files(system.file("examples", package = "knitr"),
  pattern = "\\\\.(Rnw|Rmd|html)$", full.names = TRUE)
offensive_check_files(files)
}
