% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_string.R
\name{offensive_check_string}
\alias{offensive_check_string}
\title{Checks for offensive words in strings}
\usage{
offensive_check_string(text, n_output = 2, words_add = character(),
  words_ignore = character())
}
\arguments{
\item{text}{A string of words.}

\item{n_output}{Number of surrounding words to display.}

\item{words_add}{Character. Words to look for.}

\item{words_ignore}{Character. Words to ignore.}
}
\value{
Displays offensive words and their surrounding words.
}
\description{
Checks for offensive words in strings
}
\examples{
text <- "This will be hard to asses, butt I think we will manage."
offensive_check_string(text)

offensive_check_string(text, n_output = 4)

offensive_check_string(text, words_add = "will")

offensive_check_string(text, words_ignore = "butt")
}
