% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_token.R
\name{offensive_check_token}
\alias{offensive_check_token}
\title{Checks for offensive words in tokenized vector}
\usage{
offensive_check_token(text, words_add = character(),
  words_ignore = character())
}
\arguments{
\item{text}{A vector of tokenized words.}

\item{words_add}{Character. Words to look for.}

\item{words_ignore}{Character. Words to ignore.}
}
\value{
Logical indicator.
}
\description{
Checks for offensive words in tokenized vector
}
\examples{
text <- c("this", "will", "be", "hard", "to", "asses", "butt", "i", "think",
          "we", "will", "manage")
offensive_check_token(text)
}
