% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_markdown.R
\name{parse_text}
\alias{parse_text}
\alias{parse_text_md}
\title{Text Parsers}
\usage{
parse_text_md(path, extensions = TRUE)
}
\arguments{
\item{path}{markdown file}

\item{extensions}{render markdown extensions? Passed to [commonmark][commonmark::markdown_xml]}
}
\description{
Parse text from various formats and return a data frame with text lines
and position in the source document.
}
